Clean up components section and add yaml example
