+++ vizro-core/docs/pages/user_guides/components.md
@@ -149,6 +149,8 @@ Based on the provided examples from Dash, the [`Card`][vizro.models.Card] model 
               #### Header level 4 <h4>
             type: card
           - text: |
+              ### Paragraphs
+
               Commodi repudiandae consequuntur voluptatum laborum numquam blanditiis harum quisquam eius sed odit.
 
               Fugiat iusto fuga praesentium option, eaque rerum! Provident similique accusantium nemo autem.
@@ -156,28 +158,30 @@ Based on the provided examples from Dash, the [`Card`][vizro.models.Card] model 
               Obcaecati tenetur iure eius earum ut molestias architecto voluptate aliquam nihil, eveniet aliquid.
 
               Culpa officia aut! Impedit sit sunt quaerat, odit, tenetur error, harum nesciunt ipsum debitis quas.
-            title: Paragraphs
             type: card
           - text: |
+              ### Block Quotes
+
               >
               > A block quote is a long quotation, indented to create a separate block of text.
               >
-            title: Block Quotes
             type: card
           - text: |
+              ### Lists
+
               * Item A
                 * Sub Item 1
                 * Sub Item 2
               * Item B
-            title: Lists
             type: card
           - text: |
+              ### Emphasis
+
               This word will be *italic*
 
               This word will be **bold**
 
               This word will be _**bold and italic**_
-            title: Emphasis
             type: card
           title: Customizing Text
         ```
