+++ vizro-core/examples/demo/app.py
@@ -573,4 +573,4 @@ dashboard = vm.Dashboard(
 )
 
 if __name__ == "__main__":
-    Vizro().build(dashboard).run()
+    Vizro().build(dashboard).run(port=8051)
